add .env

+++ primerComponente/views.py
@@ -53,10 +53,14 @@ class PrimerViewDetail(APIView):
 
     def delete(self, request, pk, format=None):
         idResponse = self.get_object(pk = pk)
-       
+
         if idResponse != 404:
-            idResponse.delete()
-            return Response({'message': 'usuario eliminado correctamente'})
+            serializer = PrimerTablaSerializers(idResponse, data = request.data ,context={'request': request})
+            if serializer.is_valid():
+                idResponse.delete()
+                return Response({'message': 'usuario eliminado correctamente'})
+            else:
+                return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
         else:
             return Response('ID no encontrado', status =  status.HTTP_404_NOT_FOUND) 
-      
+
